refactor: separate lib.rs
