+++ moqt-server-sample/src/main.rs
@@ -1,6 +1,6 @@
 use anyhow::{Ok, Result};
 use clap::Parser;
-use moqt_server::{constants::UnderlayType, MOQTConfig, MOQT};
+use moqt_server::{constants::UnderlayType, MOQTConfig, MOQTServer};
 
 #[derive(Debug, Parser)]
 struct Arg {
@@ -18,10 +18,8 @@ async fn main() -> Result<()> {
     config.cert_path = "./moqt-server-sample/keys/cert.pem".to_string();
     config.log_level = args.log_level.to_string();
 
-    // config.auth_callback = Some(|track_name: String, auth_payload: String, auth_callback_type: AuthCallbackType| {Ok(())});
-
-    let moqt = MOQT::new(config);
-    moqt.start().await?;
+    let moqt_server = MOQTServer::new(config);
+    moqt_server.start().await?;
 
     Ok(())
 }
