+++ moqt-server/src/modules.rs
@@ -2,6 +2,8 @@ pub mod buffer_manager;
 pub(crate) mod logging;
 pub(crate) mod message_handlers;
 pub(crate) mod moqt_client;
+pub(crate) mod moqt_config;
 pub(crate) mod object_cache_storage;
 pub(crate) mod pubsub_relation_manager;
 pub(crate) mod send_stream_dispatcher;
+pub(crate) mod server_processes;
