+++ moqt-server/src/modules/message_handlers/control_message_handler/handlers/subscribe_ok_handler.rs
@@ -48,19 +48,19 @@ pub(crate) async fn subscribe_ok_handler(
                     subscribe_ok_message.largest_object_id(),
                     subscribe_ok_message.subscribe_parameters().clone(),
                 );
-                let relaying_subscribe_ok_message: Box<dyn MOQTPayload> =
+                let forwarding_subscribe_ok_message: Box<dyn MOQTPayload> =
                     Box::new(message_payload.clone());
 
                 send_stream_dispatcher_repository
                     .transfer_message_to_send_stream_thread(
                         *downstream_session_id,
-                        relaying_subscribe_ok_message,
+                        forwarding_subscribe_ok_message,
                         StreamDirection::Bi,
                     )
                     .await?;
 
                 tracing::debug!(
-                    "message: {:#?} is sent to relay handler for client {:?}",
+                    "message: {:#?} is sent to forward handler for client {:?}",
                     message_payload,
                     downstream_session_id
                 );
@@ -91,23 +91,28 @@ pub(crate) async fn subscribe_ok_handler(
 #[cfg(test)]
 mod success {
     use super::subscribe_ok_handler;
-    use crate::modules::pubsub_relation_manager::{
-        commands::PubSubRelationCommand, manager::pubsub_relation_manager,
-        wrapper::PubSubRelationManagerWrapper,
+    use crate::modules::{
+        moqt_client::MOQTClient,
+        pubsub_relation_manager::{
+            commands::PubSubRelationCommand, manager::pubsub_relation_manager,
+            wrapper::PubSubRelationManagerWrapper,
+        },
+        send_stream_dispatcher::{
+            send_stream_dispatcher, SendStreamDispatchCommand, SendStreamDispatcher,
+        },
     };
-    use crate::modules::send_stream_dispatcher::{
-        send_stream_dispatcher, SendStreamDispatchCommand, SendStreamDispatcher,
+    use moqt_core::{
+        constants::StreamDirection,
+        messages::{
+            control_messages::{
+                subscribe::{FilterType, GroupOrder},
+                subscribe_ok::SubscribeOk,
+                version_specific_parameters::{AuthorizationInfo, VersionSpecificParameter},
+            },
+            moqt_payload::MOQTPayload,
+        },
+        pubsub_relation_manager_repository::PubSubRelationManagerRepository,
     };
-    use crate::MOQTClient;
-    use moqt_core::constants::StreamDirection;
-    use moqt_core::messages::control_messages::subscribe::{FilterType, GroupOrder};
-    use moqt_core::messages::control_messages::version_specific_parameters::{
-        AuthorizationInfo, VersionSpecificParameter,
-    };
-    use moqt_core::messages::{
-        control_messages::subscribe_ok::SubscribeOk, moqt_payload::MOQTPayload,
-    };
-    use moqt_core::pubsub_relation_manager_repository::PubSubRelationManagerRepository;
     use std::sync::Arc;
     use tokio::sync::mpsc;
 
@@ -147,7 +152,7 @@ mod success {
 
         // Generate client
         let upstream_session_id = 1;
-        let client = MOQTClient::new(upstream_session_id);
+        let client = MOQTClient::new_without_senders(upstream_session_id);
 
         // Generate PubSubRelationManagerWrapper
         let (track_namespace_tx, mut track_namespace_rx) =
@@ -216,12 +221,12 @@ mod success {
         let mut send_stream_dispatcher: SendStreamDispatcher =
             SendStreamDispatcher::new(send_stream_tx.clone());
 
-        let (uni_relay_tx, _) = mpsc::channel::<Arc<Box<dyn MOQTPayload>>>(1024);
+        let (message_tx, _) = mpsc::channel::<Arc<Box<dyn MOQTPayload>>>(1024);
         let _ = send_stream_tx
             .send(SendStreamDispatchCommand::Set {
                 session_id: downstream_session_id,
                 stream_direction: StreamDirection::Bi,
-                sender: uni_relay_tx,
+                sender: message_tx,
             })
             .await;
 
@@ -240,29 +245,34 @@ mod success {
 
 #[cfg(test)]
 mod failure {
-    use crate::modules::message_handlers::control_message_handler::handlers::subscribe_ok_handler::subscribe_ok_handler;
-    use crate::modules::pubsub_relation_manager::{
-        commands::PubSubRelationCommand, manager::pubsub_relation_manager,
-        wrapper::PubSubRelationManagerWrapper,
+    use super::subscribe_ok_handler;
+    use crate::modules::{
+        moqt_client::MOQTClient,
+        pubsub_relation_manager::{
+            commands::PubSubRelationCommand, manager::pubsub_relation_manager,
+            wrapper::PubSubRelationManagerWrapper,
+        },
+        send_stream_dispatcher::{
+            send_stream_dispatcher, SendStreamDispatchCommand, SendStreamDispatcher,
+        },
     };
-    use crate::modules::send_stream_dispatcher::{
-        send_stream_dispatcher, SendStreamDispatchCommand, SendStreamDispatcher,
+    use moqt_core::{
+        constants::StreamDirection,
+        messages::{
+            control_messages::{
+                subscribe::{FilterType, GroupOrder},
+                subscribe_ok::SubscribeOk,
+                version_specific_parameters::{AuthorizationInfo, VersionSpecificParameter},
+            },
+            moqt_payload::MOQTPayload,
+        },
+        pubsub_relation_manager_repository::PubSubRelationManagerRepository,
     };
-    use crate::MOQTClient;
-    use moqt_core::constants::StreamDirection;
-    use moqt_core::messages::control_messages::subscribe::{FilterType, GroupOrder};
-    use moqt_core::messages::control_messages::version_specific_parameters::{
-        AuthorizationInfo, VersionSpecificParameter,
-    };
-    use moqt_core::messages::{
-        control_messages::subscribe_ok::SubscribeOk, moqt_payload::MOQTPayload,
-    };
-    use moqt_core::pubsub_relation_manager_repository::PubSubRelationManagerRepository;
     use std::sync::Arc;
     use tokio::sync::mpsc;
 
     #[tokio::test]
-    async fn relay_fail() {
+    async fn forward_fail() {
         // Generate SUBSCRIBE_OK message
         let downstream_subscribe_id = 0;
         let track_alias = 0;
@@ -297,7 +307,7 @@ mod failure {
 
         // Generate client
         let upstream_session_id = 1;
-        let client = MOQTClient::new(upstream_session_id);
+        let client = MOQTClient::new_without_senders(upstream_session_id);
 
         // Generate PubSubRelationManagerWrapper
         let (track_namespace_tx, mut track_namespace_rx) =
@@ -413,7 +423,7 @@ mod failure {
 
         // Generate client
         let upstream_session_id = 1;
-        let client = MOQTClient::new(upstream_session_id);
+        let client = MOQTClient::new_without_senders(upstream_session_id);
 
         // Generate PubSubRelationManagerWrapper
         let (track_namespace_tx, mut track_namespace_rx) =
@@ -454,12 +464,12 @@ mod failure {
         let mut send_stream_dispatcher: SendStreamDispatcher =
             SendStreamDispatcher::new(send_stream_tx.clone());
 
-        let (uni_relay_tx, _) = mpsc::channel::<Arc<Box<dyn MOQTPayload>>>(1024);
+        let (message_tx, _) = mpsc::channel::<Arc<Box<dyn MOQTPayload>>>(1024);
         let _ = send_stream_tx
             .send(SendStreamDispatchCommand::Set {
                 session_id: downstream_session_id,
                 stream_direction: StreamDirection::Bi,
-                sender: uni_relay_tx,
+                sender: message_tx,
             })
             .await;
 
@@ -511,7 +521,7 @@ mod failure {
 
         // Generate client
         let upstream_session_id = 1;
-        let client = MOQTClient::new(upstream_session_id);
+        let client = MOQTClient::new_without_senders(upstream_session_id);
 
         // Generate PubSubRelationManagerWrapper
         let (track_namespace_tx, mut track_namespace_rx) =
@@ -584,12 +594,12 @@ mod failure {
         let mut send_stream_dispatcher: SendStreamDispatcher =
             SendStreamDispatcher::new(send_stream_tx.clone());
 
-        let (uni_relay_tx, _) = mpsc::channel::<Arc<Box<dyn MOQTPayload>>>(1024);
+        let (message_tx, _) = mpsc::channel::<Arc<Box<dyn MOQTPayload>>>(1024);
         let _ = send_stream_tx
             .send(SendStreamDispatchCommand::Set {
                 session_id: downstream_session_id,
                 stream_direction: StreamDirection::Bi,
-                sender: uni_relay_tx,
+                sender: message_tx,
             })
             .await;
 
@@ -641,7 +651,7 @@ mod failure {
 
         // Generate client
         let upstream_session_id = 1;
-        let client = MOQTClient::new(upstream_session_id);
+        let client = MOQTClient::new_without_senders(upstream_session_id);
 
         // Generate PubSubRelationManagerWrapper
         let (track_namespace_tx, mut track_namespace_rx) =
@@ -714,12 +724,12 @@ mod failure {
         let mut send_stream_dispatcher: SendStreamDispatcher =
             SendStreamDispatcher::new(send_stream_tx.clone());
 
-        let (uni_relay_tx, _) = mpsc::channel::<Arc<Box<dyn MOQTPayload>>>(1024);
+        let (message_tx, _) = mpsc::channel::<Arc<Box<dyn MOQTPayload>>>(1024);
         let _ = send_stream_tx
             .send(SendStreamDispatchCommand::Set {
                 session_id: downstream_session_id,
                 stream_direction: StreamDirection::Bi,
-                sender: uni_relay_tx,
+                sender: message_tx,
             })
             .await;
