+++ moqt-server/src/modules/message_handlers/control_message_handler/server_processes/subscribe_message.rs
@@ -22,7 +22,7 @@ pub(crate) async fn process_subscribe_message(
     pubsub_relation_manager_repository: &mut dyn PubSubRelationManagerRepository,
     send_stream_dispatcher_repository: &mut dyn SendStreamDispatcherRepository,
     object_cache_storage: &mut ObjectCacheStorageWrapper,
-    open_downstream_subscription_txes: Arc<Mutex<HashMap<usize, SenderToOpenSubscription>>>,
+    open_downstream_stream_or_datagram_txes: Arc<Mutex<HashMap<usize, SenderToOpenSubscription>>>,
 ) -> Result<Option<SubscribeError>> {
     let subscribe_request_message = match Subscribe::depacketize(payload_buf) {
         Ok(subscribe_request_message) => subscribe_request_message,
@@ -38,7 +38,7 @@ pub(crate) async fn process_subscribe_message(
         pubsub_relation_manager_repository,
         send_stream_dispatcher_repository,
         object_cache_storage,
-        open_downstream_subscription_txes,
+        open_downstream_stream_or_datagram_txes,
     )
     .await;
