+++ moqt-server/src/modules/moqt_client.rs
@@ -1,6 +1,7 @@
+use super::server_processes::senders::Senders;
 use anyhow::{bail, Ok, Result};
-
 use moqt_core::messages::control_messages::setup_parameters::RoleCase;
+use std::sync::Arc;
 
 #[derive(Debug, PartialEq, Eq, Clone, Copy)]
 pub enum MOQTClientStatus {
@@ -13,14 +14,17 @@ pub struct MOQTClient {
     id: usize,
     status: MOQTClientStatus,
     role: Option<RoleCase>,
+    senders: Option<Arc<Senders>>, // TODO: remove Option
 }
 
 impl MOQTClient {
-    pub fn new(id: usize) -> Self {
+    pub fn new(id: usize, senders: Senders) -> Self {
+        let senders = Arc::new(senders);
         MOQTClient {
             id,
             status: MOQTClientStatus::Connected,
             role: None,
+            senders: Some(senders),
         }
     }
     pub fn id(&self) -> usize {
@@ -43,5 +47,19 @@ impl MOQTClient {
         self.role = Some(new_role);
 
         Ok(())
+    }
+
+    pub fn senders(&self) -> Arc<Senders> {
+        self.senders.clone().unwrap()
+    }
+
+    #[cfg(test)]
+    pub fn new_without_senders(id: usize) -> Self {
+        MOQTClient {
+            id,
+            status: MOQTClientStatus::Connected,
+            role: None,
+            senders: None,
+        }
     }
 }
