+++ moqt-server/src/modules/moqt_config.rs
@@ -1,0 +1,30 @@
+use moqt_core::constants::UnderlayType;
+
+pub struct MOQTConfig {
+    pub port: u16,
+    pub cert_path: String,
+    pub key_path: String,
+    pub keep_alive_interval_sec: u64,
+    pub underlay: UnderlayType,
+    pub log_level: String,
+}
+
+impl Default for MOQTConfig {
+    fn default() -> Self {
+        MOQTConfig::new()
+    }
+}
+
+impl MOQTConfig {
+    // TODO: use getter/setter
+    pub fn new() -> MOQTConfig {
+        MOQTConfig {
+            port: 4433,
+            cert_path: "./cert.pem".to_string(),
+            key_path: "./key.pem".to_string(),
+            keep_alive_interval_sec: 3,
+            underlay: UnderlayType::Both,
+            log_level: "DEBUG".to_string(),
+        }
+    }
+}
