+++ moqt-server/src/modules/server_processes/stream_and_datagram/bi_directional_stream/receiver.rs
@@ -1,0 +1,98 @@
+use crate::modules::{
+    buffer_manager::{request_buffer, BufferCommand},
+    message_handlers::control_message_handler::{control_message_handler, MessageProcessResult},
+    moqt_client::MOQTClient,
+    pubsub_relation_manager::wrapper::PubSubRelationManagerWrapper,
+    send_stream_dispatcher::SendStreamDispatcher,
+};
+use anyhow::Result;
+use bytes::BytesMut;
+use moqt_core::constants::UnderlayType;
+use std::sync::Arc;
+use tokio::sync::Mutex;
+use tracing::{self};
+
+use super::stream::BiStream;
+
+pub(crate) async fn handle_bi_recv_stream(
+    stream: &mut BiStream,
+    client: Arc<Mutex<MOQTClient>>,
+) -> Result<()> {
+    let senders = client.lock().await.senders();
+    let mut buffer = vec![0; 65536].into_boxed_slice();
+
+    let stable_id = stream.stable_id();
+    let stream_id = stream.stream_id();
+    let recv_stream = &mut stream.recv_stream;
+    let shared_send_stream = &mut stream.shared_send_stream;
+
+    let mut pubsub_relation_manager =
+        PubSubRelationManagerWrapper::new(senders.pubsub_relation_tx().clone());
+    let mut send_stream_dispatcher = SendStreamDispatcher::new(senders.send_stream_tx().clone());
+
+    let mut object_cache_storage =
+        crate::modules::object_cache_storage::ObjectCacheStorageWrapper::new(
+            senders.object_cache_tx().clone(),
+        );
+
+    loop {
+        let bytes_read = match recv_stream.read(&mut buffer).await? {
+            Some(bytes_read) => bytes_read,
+            None => break,
+        };
+
+        tracing::debug!("bytes_read: {}", bytes_read);
+
+        let read_buf = BytesMut::from(&buffer[..bytes_read]);
+        let buf = request_buffer(senders.buffer_tx().clone(), stable_id, stream_id).await;
+        let mut buf = buf.lock().await;
+        buf.extend_from_slice(&read_buf);
+
+        let message_result: MessageProcessResult;
+        {
+            let mut client = client.lock().await;
+            // TODO: Move the implementation of control_message_handler to the server side since it is only used by the server
+            message_result = control_message_handler(
+                &mut buf,
+                UnderlayType::WebTransport,
+                &mut client,
+                senders.open_downstream_stream_or_datagram_txes().clone(),
+                &mut pubsub_relation_manager,
+                &mut send_stream_dispatcher,
+                &mut object_cache_storage,
+            )
+            .await;
+        }
+
+        tracing::debug!("message_result: {:?}", message_result);
+
+        match message_result {
+            MessageProcessResult::Success(buf) => {
+                let mut shared_send_stream = shared_send_stream.lock().await;
+                shared_send_stream.write_all(&buf).await?;
+
+                tracing::info!("Message is sent.");
+                tracing::debug!("sent message: {:x?}", buf.to_vec());
+            }
+            MessageProcessResult::SuccessWithoutResponse => {}
+            MessageProcessResult::Failure(code, message) => {
+                senders
+                    .close_session_tx()
+                    .send((u8::from(code) as u64, message))
+                    .await?;
+                break;
+            }
+            MessageProcessResult::Fragment => (),
+        };
+    }
+
+    senders
+        .buffer_tx()
+        .send(BufferCommand::ReleaseStream {
+            session_id: stable_id,
+            stream_id,
+        })
+        .await?;
+
+    Ok(())
+}
